refactor: Move 4 misplaced utils modules to canonical locations (#873)
Relocate single-consumer modules from utils/numerical/ to their
logical homes. Old import paths remain as deprecation shims
(removal at v0.21.0, per 3-version lifecycle policy).

Moves:
- gfdm_strategies.py (2,618 lines) -> alg/numerical/gfdm_components/
- anderson_acceleration.py (519 lines) -> alg/numerical/coupling/
- sdf_utils.py (550 lines) -> geometry/implicit/
- mesh_distances.py (213 lines) -> geometry/meshes/

All direct importers updated to canonical paths.
utils/numerical/__init__.py re-exports from new locations.
